Introduce the Replit authentication (#13)
This change adds the Replit authentication. This uses the VeNCrypt Plain
authentication under the covers, and provides the Goval Handshake v5
token as username, and it is validated for being not expired AND being
issued for the Repl that it's running as.

The password will be forwarded to TigerVNC using the VncAuth (the broken
one that uses DES) if it claims to support passwords, or ignored
otherwise.